share logic between typeahead and next button
